# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Betera test task in Gin Framework",
        "title": "Swagger Documentation",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/nasa/{date}": {
            "get": {
                "description": "Endpoint for getting the APOD with exact date",
                "produces": [
                    "application/json"
                ],
                "summary": "The exact APOD",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/nasa.Metadata"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "nasa.Metadata": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "explanation": {
                    "type": "string"
                },
                "media_type": {
                    "type": "string"
                },
                "service_version": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}